commit colorbars example. Fix set_palette.

+++ libmindset_gfx/libmindset_gfx.c
@@ -288,7 +288,7 @@ void mindset_gfx_set_display_interrupt(unsigned char line, unsigned char marker)
  * U - Unused
  *
  */
-void mindset_gfx_set_palette(unsigned char border, unsigned short num, unsigned short offset, unsigned long* palette)
+void mindset_gfx_set_palette(unsigned char border, unsigned short num, unsigned short offset, unsigned short* palette)
 {
   union REGS regs;
   struct SREGS sregs;
